analytics dashboard.

+++ src/main/java/com/sabi/suppliers/core/models/response/SupplierDashbaordResponseDto.java
@@ -3,7 +3,9 @@ package com.sabi.suppliers.core.models.response;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -13,10 +15,10 @@ public class SupplierDashbaordResponseDto {
     private int completedOrder;
     private int cancelledOrder;
     private int ongoingDelivery;
-    private double totalSales;
+    private BigDecimal totalSales;
     private int warehouses;
     private int products;
-    private double outstandingPayment;
+    private BigDecimal outstandingPayment;
     private Boolean isActive;
     private LocalDateTime createdDate;
     private LocalDateTime updatedDate;
